Resize or create filesystem
This makes the assumption that the filesystem only needs creating if
the LV is fresh. This does mean it is possible to get into a state where
the LV has been created, but with no filesystem on it - if the filesystem
was not created successfully originally).

Not sure whether to consider this a bug though. Potentially could check for
a filesystem on the device before doing the mkfs.